+++ nexus/db-model/src/vmm_state.rs
@@ -145,28 +145,7 @@ impl From<sled_agent_types::instance::VmmState> for VmmState {
 
 impl From<VmmState> for omicron_common::api::external::InstanceState {
     fn from(value: VmmState) -> Self {
-        use omicron_common::api::external::InstanceState as Output;
-
-        match value {
-            // An instance with a VMM which is in the `Creating` state maps to
-            // `InstanceState::Starting`, rather than `InstanceState::Creating`.
-            // If we are still creating the VMM, this is because we are
-            // attempting to *start* the instance; instances may be created
-            // without creating a VMM to run them, and then started later.
-            VmmState::Creating | VmmState::Starting => Output::Starting,
-            VmmState::Running => Output::Running,
-            VmmState::Stopping => Output::Stopping,
-            // `SagaUnwound` should map to `Stopped` so that an `instance_view`
-            // API call that produces an instance with an unwound VMM will appear to
-            // be `Stopped`. This is because instances with unwound VMMs can
-            // be started by a subsequent instance-start saga, just like
-            // instances whose internal state actually is `Stopped`.
-            VmmState::Stopped | VmmState::SagaUnwound => Output::Stopped,
-            VmmState::Rebooting => Output::Rebooting,
-            VmmState::Migrating => Output::Migrating,
-            VmmState::Failed => Output::Failed,
-            VmmState::Destroyed => Output::Destroyed,
-        }
+        value.to_nexus_state().into()
     }
 }
